Move ETH back into SDK
