+++ packages/sdk/src/constants/config.ts
@@ -1,0 +1,9 @@
+/**
+ * Default SDK configuration values
+ */
+export const DEFAULT_VERBS_CONFIG = {
+  lend: {
+    /** Default slippage tolerance for lending operations (in basis points: 50 = 0.5%) */
+    defaultSlippage: 50,
+  },
+} as const
