+++ packages/sdk/src/utils/config.ts
@@ -1,16 +1,6 @@
 import type { LendMarketConfig, LendMarketId } from '@/types/lend.js'
 
 /**
- * Default SDK configuration values
- */
-export const DEFAULT_VERBS_CONFIG = {
-  lend: {
-    /** Default slippage tolerance for lending operations (in basis points: 50 = 0.5%) */
-    defaultSlippage: 50,
-  },
-} as const
-
-/**
  * Find a lend market configuration from a market allowlist
  * @param marketAllowlist - List of allowed markets
  * @param lendMarketId - Market identifier to find
