+++ packages/sdk/src/verbs.ts
@@ -1,9 +1,9 @@
+import { DEFAULT_VERBS_CONFIG } from '@/constants/config.js'
 import { LendProviderMorpho } from '@/lend/index.js'
 import { VerbsLendNamespace } from '@/lend/namespaces/VerbsLendNamespace.js'
 import { ChainManager } from '@/services/ChainManager.js'
 import type { LendProvider } from '@/types/lend.js'
 import type { VerbsConfig } from '@/types/verbs.js'
-import { DEFAULT_VERBS_CONFIG } from '@/utils/config.js'
 import type { SmartWalletProvider } from '@/wallet/providers/base/SmartWalletProvider.js'
 import { DefaultSmartWalletProvider } from '@/wallet/providers/DefaultSmartWalletProvider.js'
 import type {
